Created 3 methods in signInActivity.java class to reset user’s password: forgotPassword(), showForgotPasswordDialog(), and beginPasswordReset().
I added an alert dialog theme to styles.xml and colors.xml because our current theme made the buttons white, which was hard to see. I just put a basic red for negative buttons and a green for positive buttons. You can change the colors to whatever you think fits the theme best.